+++ cmd/migrate_cmd.go
@@ -1,6 +1,7 @@
 package cmd
 
 import (
+	"embed"
 	"fmt"
 	"net/url"
 	"os"
@@ -11,6 +12,8 @@ import (
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 )
+
+var EmbeddedMigrations embed.FS
 
 var migrateCmd = cobra.Command{
 	Use:  "migrate",
@@ -76,11 +79,14 @@ func migrate(cmd *cobra.Command, args []string) {
 		log.Fatalf("%+v", errors.Wrap(err, "checking database connection"))
 	}
 
-	log.Debugf("Reading migrations from %s", globalConfig.DB.MigrationsPath)
-	mig, err := pop.NewFileMigrator(globalConfig.DB.MigrationsPath, db)
+	log.Debugf("Reading migrations from executable")
+	box, err := pop.NewMigrationBox(EmbeddedMigrations, db)
 	if err != nil {
 		log.Fatalf("%+v", errors.Wrap(err, "creating db migrator"))
 	}
+
+	mig := box.Migrator
+
 	log.Debugf("before status")
 
 	if log.Level == logrus.DebugLevel {