+++ internal/api/helpers.go
@@ -12,6 +12,10 @@ import (
 
 	"github.com/gofrs/uuid"
 	"github.com/pkg/errors"
+	"github.com/supabase/auth/internal/conf"
+	"github.com/supabase/auth/internal/models"
+	"github.com/supabase/auth/internal/security"
+	"github.com/supabase/auth/internal/utilities"
 )
 
 func sendJSON(w http.ResponseWriter, status int, obj interface{}) error {
@@ -75,11 +79,6 @@ func isStringInSlice(checkValue string, list []string) bool {
 	return false
 }
 
-// getBodyBytes returns a byte array of the request's Body.
-func getBodyBytes(req *http.Request) ([]byte, error) {
-	return utilities.GetBodyBytes(req)
-}
-
 type RequestParams interface {
 	AdminUserParams |
 		CreateSSOProviderParams |
@@ -100,6 +99,8 @@ type RequestParams interface {
 		VerifyFactorParams |
 		VerifyParams |
 		adminUserUpdateFactorParams |
+		adminUserDeleteParams |
+		security.GotrueRequest |
 		ChallengeFactorParams |
 		struct {
 			Email string `json:"email"`
@@ -112,7 +113,7 @@ type RequestParams interface {
 
 // retrieveRequestParams is a generic method that unmarshals the request body into the params struct provided
 func retrieveRequestParams[A RequestParams](r *http.Request, params *A) error {
-	body, err := getBodyBytes(r)
+	body, err := utilities.GetBodyBytes(r)
 	if err != nil {
 		return internalServerError("Could not read body into byte slice").WithInternalError(err)
 	}